pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jvdeploy"
dynamic = ["version"]
description = "Dockerfile generator for jvagent applications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "TrueSelph Inc.", email = "adminh@trueselph.com"},
]
keywords = ["docker", "dockerfile", "bundler", "jvagent", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
deploy = [
    "boto3>=1.28.0",
    "jinja2>=3.1.0",
]
dev = [
    "pytest>=7.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "boto3>=1.28.0",
    "jinja2>=3.1.0",
]
test = [
    "pytest>=7.0",
]

[project.scripts]
jvdeploy = "jvdeploy.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/jvdeploy"
Repository = "https://github.com/your-org/jvdeploy"
Issues = "https://github.com/your-org/jvdeploy/issues"

[tool.setuptools.dynamic]
version = {attr = "jvdeploy.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["jvdeploy*"]

[tool.setuptools.package-data]
jvdeploy = ["*.base", "templates/*.template", "templates/*.j2"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
